Create CalcVector
This program multiplies the vectors <6, 10, -3> and <-11, 0, -9>. First, it calculates and prints the cross product of the two vectors. Then, it calculates the magnitude of the resultant vector. Finally, it produces the dot product of the two vectors. One may change the numbers in the setVector method in main or uncomment the block in main to input vector elements when the program is run.