Create Apple_and_Orange
A house has an apple and an orange tree. The endpoints of the house are given where s is the starting point and t is the ending point. The apple tree is to the left of his house, and the orange tree is to its right. You can assume the trees are located on a single point, where the apple tree is at point a, and the orange tree is at point b.
When a fruit falls from its tree, it lands d units of distance from its tree of origin along the x-axis. A negative value of d means the fruit fell d units to the tree's left, and a positive value of d means it falls d units to the tree's right.
Given the value of d for m apples and n oranges, determine how many apples and oranges will fall on the house (i.e., in the inclusive range )?