Solved Tower of Hanoi using recursion
Tower of Hanoi, is a mathematical puzzle which consists of three towers  and more than one rings.
These rings are of different sizes and stacked upon in an ascending order, i.e. the smaller one sits over the larger one. There are other variations of the puzzle where the number of disks increase, but the tower count remains the same.
The mission is to move all the disks to some another tower without violating the sequence of arrangement. A few rules to be followed for Tower of Hanoi are −
1) Only one disk can be moved among the towers at any given time.
2) Only the "top" disk can be removed.
3) No large disk can sit over a small disk.